+++ src/main/java/jsclub/codefest/sdk/socket/data/receive_data/ItemData.java
@@ -15,7 +15,7 @@ public class ItemData {
     public int level;
 
     @SerializedName("durability")
-    public int durability;
+    public Long durability;
 
     @SerializedName("attributes")
     public Object attributes;
